backend/requirements.txt
googlemaps
osmnx
networkx
shapely
geopandas
pysolar
pandas
requests
numpy
python-dotenv
fastapi[standard]
uvicorn
pydantic
sqlalchemy
pymysql
cryptography